requirements.txt
streamlit>=1.28.0
streamlit-webrtc>=0.44.4
face-recognition>=1.3.0
opencv-python>=4.6.0
numpy>=1.23.5
pandas>=1.5.3
pillow>=9.4.0
python-dotenv>=1.0.0
openai>=1.0.0
groq
plotly
google-generativeai
dlib
imutils